GH-282 Extract decorators
