+++ region-connectors/shared/build.gradle.kts
@@ -1,0 +1,37 @@
+import net.ltgt.gradle.errorprone.CheckSeverity
+import net.ltgt.gradle.errorprone.errorprone
+
+plugins {
+    id("energy.eddie.java-conventions")
+}
+group = "energy.eddie"
+version = "0.0.0"
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation(project(":api"))
+
+    implementation(libs.reactor.core)
+    implementation(project(mapOf("path" to ":api")))
+    testImplementation(libs.reactor.test)
+
+    testImplementation(platform("org.junit:junit-bom:5.9.1"))
+    testImplementation("org.junit.jupiter:junit-jupiter")
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+tasks.withType<JavaCompile>().configureEach {
+    options.errorprone {
+        check("NullAway", CheckSeverity.ERROR)
+        option("NullAway:AnnotatedPackages", "energy.eddie")
+
+        // disable warnings for generated classes
+        option("NullAway:TreatGeneratedAsUnannotated", true)
+    }
+}
