+++ region-connectors/shared/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module energy.eddie.region.connector.shared {
+    requires energy.eddie.api;
+    requires reactor.core;
+    exports energy.eddie.regionconnector.shared.permission.requests.decorators;
+}
