+++ shared.iml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="SonarLintModuleSettings">
+    <option name="uniqueId" value="983b67d6-b6b0-4664-af6b-06ee8b231c0b" />
+  </component>
+</module>
